+++ tests/testsuite/registry.rs
@@ -1078,145 +1078,6 @@ fn dev_dependency_not_used(cargo: fn(&Project, &str) -> Execs) {
 }
 
 #[cargo_test]
-fn login_with_no_cargo_dir() {
-    // Create a config in the root directory because `login` requires the
-    // index to be updated, and we don't want to hit crates.io.
-    let registry = registry::init();
-    fs::rename(paths::home().join(".cargo"), paths::root().join(".cargo")).unwrap();
-    paths::home().rm_rf();
-    cargo_process("login foo -v")
-        .replace_crates_io(registry.index_url())
-        .run();
-    let credentials = fs::read_to_string(paths::home().join(".cargo/credentials")).unwrap();
-    assert_eq!(credentials, "[registry]\ntoken = \"foo\"\n");
-}
-
-#[cargo_test]
-fn login_with_differently_sized_token() {
-    // Verify that the configuration file gets properly truncated.
-    let registry = registry::init();
-    let credentials = paths::home().join(".cargo/credentials");
-    fs::remove_file(&credentials).unwrap();
-    cargo_process("login lmaolmaolmao -v")
-        .replace_crates_io(registry.index_url())
-        .run();
-    cargo_process("login lmao -v")
-        .replace_crates_io(registry.index_url())
-        .run();
-    cargo_process("login lmaolmaolmao -v")
-        .replace_crates_io(registry.index_url())
-        .run();
-    let credentials = fs::read_to_string(&credentials).unwrap();
-    assert_eq!(credentials, "[registry]\ntoken = \"lmaolmaolmao\"\n");
-}
-
-#[cargo_test]
-fn login_with_token_on_stdin() {
-    let registry = registry::init();
-    let credentials = paths::home().join(".cargo/credentials");
-    fs::remove_file(&credentials).unwrap();
-    cargo_process("login lmao -v")
-        .replace_crates_io(registry.index_url())
-        .run();
-    cargo_process("login")
-        .replace_crates_io(registry.index_url())
-        .with_stdout("please paste the token found on [..]/me below")
-        .with_stdin("some token")
-        .run();
-    let credentials = fs::read_to_string(&credentials).unwrap();
-    assert_eq!(credentials, "[registry]\ntoken = \"some token\"\n");
-}
-
-#[cargo_test]
-fn login_with_asymmetric_token_and_subject_on_stdin() {
-    let registry = registry::init();
-    let credentials = paths::home().join(".cargo/credentials");
-    fs::remove_file(&credentials).unwrap();
-    cargo_process("login --key-subject=foo --secret-key -v -Z registry-auth")
-        .masquerade_as_nightly_cargo(&["registry-auth"])
-        .replace_crates_io(registry.index_url())
-        .with_stdout(
-            "\
-        please paste the API secret key below
-k3.public.AmDwjlyf8jAV3gm5Z7Kz9xAOcsKslt_Vwp5v-emjFzBHLCtcANzTaVEghTNEMj9PkQ",
-        )
-        .with_stdin("k3.secret.fNYVuMvBgOlljt9TDohnaYLblghqaHoQquVZwgR6X12cBFHZLFsaU3q7X3k1Zn36")
-        .run();
-    let credentials = fs::read_to_string(&credentials).unwrap();
-    assert!(credentials.starts_with("[registry]\n"));
-    assert!(credentials.contains("secret-key-subject = \"foo\"\n"));
-    assert!(credentials.contains("secret-key = \"k3.secret.fNYVuMvBgOlljt9TDohnaYLblghqaHoQquVZwgR6X12cBFHZLFsaU3q7X3k1Zn36\"\n"));
-}
-
-#[cargo_test]
-fn login_with_asymmetric_token_on_stdin() {
-    let registry = registry::init();
-    let credentials = paths::home().join(".cargo/credentials");
-    fs::remove_file(&credentials).unwrap();
-    cargo_process("login --secret-key -v -Z registry-auth")
-        .masquerade_as_nightly_cargo(&["registry-auth"])
-        .replace_crates_io(registry.index_url())
-        .with_stdout(
-            "\
-    please paste the API secret key below
-k3.public.AmDwjlyf8jAV3gm5Z7Kz9xAOcsKslt_Vwp5v-emjFzBHLCtcANzTaVEghTNEMj9PkQ",
-        )
-        .with_stdin("k3.secret.fNYVuMvBgOlljt9TDohnaYLblghqaHoQquVZwgR6X12cBFHZLFsaU3q7X3k1Zn36")
-        .run();
-    let credentials = fs::read_to_string(&credentials).unwrap();
-    assert_eq!(credentials, "[registry]\nsecret-key = \"k3.secret.fNYVuMvBgOlljt9TDohnaYLblghqaHoQquVZwgR6X12cBFHZLFsaU3q7X3k1Zn36\"\n");
-}
-
-#[cargo_test]
-fn login_with_asymmetric_key_subject_without_key() {
-    let registry = registry::init();
-    let credentials = paths::home().join(".cargo/credentials");
-    fs::remove_file(&credentials).unwrap();
-    cargo_process("login --key-subject=foo -Z registry-auth")
-        .masquerade_as_nightly_cargo(&["registry-auth"])
-        .replace_crates_io(registry.index_url())
-        .with_stderr_contains("error: need a secret_key to set a key_subject")
-        .with_status(101)
-        .run();
-
-    // ok so add a secret_key to the credentials
-    cargo_process("login --secret-key -v -Z registry-auth")
-        .masquerade_as_nightly_cargo(&["registry-auth"])
-        .replace_crates_io(registry.index_url())
-        .with_stdout(
-            "please paste the API secret key below
-k3.public.AmDwjlyf8jAV3gm5Z7Kz9xAOcsKslt_Vwp5v-emjFzBHLCtcANzTaVEghTNEMj9PkQ",
-        )
-        .with_stdin("k3.secret.fNYVuMvBgOlljt9TDohnaYLblghqaHoQquVZwgR6X12cBFHZLFsaU3q7X3k1Zn36")
-        .run();
-
-    // and then it shuld work
-    cargo_process("login --key-subject=foo -Z registry-auth")
-        .masquerade_as_nightly_cargo(&["registry-auth"])
-        .replace_crates_io(registry.index_url())
-        .run();
-
-    let credentials = fs::read_to_string(&credentials).unwrap();
-    assert!(credentials.starts_with("[registry]\n"));
-    assert!(credentials.contains("secret-key-subject = \"foo\"\n"));
-    assert!(credentials.contains("secret-key = \"k3.secret.fNYVuMvBgOlljt9TDohnaYLblghqaHoQquVZwgR6X12cBFHZLFsaU3q7X3k1Zn36\"\n"));
-}
-
-#[cargo_test]
-fn login_with_generate_asymmetric_token() {
-    let registry = registry::init();
-    let credentials = paths::home().join(".cargo/credentials");
-    fs::remove_file(&credentials).unwrap();
-    cargo_process("login --generate-keypair -Z registry-auth")
-        .masquerade_as_nightly_cargo(&["registry-auth"])
-        .replace_crates_io(registry.index_url())
-        .with_stdout("k3.public.[..]")
-        .run();
-    let credentials = fs::read_to_string(&credentials).unwrap();
-    assert!(credentials.contains("secret-key = \"k3.secret."));
-}
-
-#[cargo_test]
 fn bad_license_file_http() {
     let registry = setup_http();
     bad_license_file(cargo_http, &registry);
